refactor: split MainArea component into smaller focused components
- Extract FileUploadZone component (88 lines) for file upload logic
- Extract DownloadSection component (26 lines) for download display
- Refactor MainArea component (90 lines) to focus on main processing logic
- Update tests to reflect new component structure
- Improve code maintainability by separating concerns

This change reduces MainArea from 165 lines to 90 lines while maintaining
all existing functionality through proper component composition.
